refactor: Add missing subpackages
