+++ app/src/main/java/com/monstarlab/features/resources/data/storage/ResourcePreferenceStore.kt
@@ -3,7 +3,7 @@ package com.monstarlab.features.resources.data.storage
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import com.monstarlab.core.persistence.SharedPreferenceDataStore
-import com.monstarlab.features.resources.domain.Resource
+import com.monstarlab.features.resources.domain.model.Resource
 import javax.inject.Inject
 
 class ResourcePreferenceStore @Inject constructor(
