+++ app/src/main/java/com/monstarlab/features/resources/domain/model/Resource.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.features.resources.domain
+package com.monstarlab.features.resources.domain.model
 
 import kotlinx.serialization.Serializable
