+++ app/src/main/java/com/monstarlab/features/resources/domain/repository/ResourceRepository.kt
@@ -1,9 +1,10 @@
-package com.monstarlab.features.resources.domain
+package com.monstarlab.features.resources.domain.repository
 
 import com.monstarlab.core.persistence.Repository
 import com.monstarlab.features.resources.data.api.ResourcesApi
 import com.monstarlab.features.resources.data.api.dtos.toResource
 import com.monstarlab.features.resources.data.storage.ResourcePreferenceStore
+import com.monstarlab.features.resources.domain.model.Resource
 import javax.inject.Inject
 
 class ResourceRepository @Inject constructor(
