+++ app/src/main/java/com/monstarlab/features/resources/domain/usecase/GetResourcesUseCase.kt
@@ -1,6 +1,7 @@
-package com.monstarlab.features.resources.domain
+package com.monstarlab.features.resources.domain.usecase
 
 import com.monstarlab.core.extensions.suspendRunCatching
+import com.monstarlab.features.resources.domain.repository.ResourceRepository
 import kotlinx.coroutines.delay
 import javax.inject.Inject
