+++ app/src/main/java/com/monstarlab/features/resources/ui/ResourcesContract.kt
@@ -1,7 +1,7 @@
 package com.monstarlab.features.resources.ui
 
 import com.monstarlab.core.error.ErrorModel
-import com.monstarlab.features.resources.domain.Resource
+import com.monstarlab.features.resources.domain.model.Resource
 
 /**
  * UI State that represents ResourcesScreen
