+++ app/src/main/java/com/monstarlab/features/user/data/storage/UserPreferenceStore.kt
@@ -3,7 +3,7 @@ package com.monstarlab.features.user.data.storage
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import com.monstarlab.core.persistence.SingleSharedPreferenceDataStore
-import com.monstarlab.features.user.domain.User
+import com.monstarlab.features.user.domain.model.User
 import javax.inject.Inject
 
 class UserPreferenceStore @Inject constructor(
